Simplify reviews render method
