+++ assets/js/blocks/reviews/all-reviews/edit.js
@@ -62,7 +62,6 @@ const AllReviewsEditor = ( { attributes, setAttributes } ) => {
 			{ getInspectorControls() }
 			<EditorContainerBlock
 				attributes={ attributes }
-				className="wc-block-all-reviews"
 				icon={
 					<Icon
 						icon={ discussion }
