+++ assets/js/blocks/reviews/reviews-by-product/edit.js
@@ -162,7 +162,6 @@ const ReviewsByProductEditor = ( {
 			{ getInspectorControls() }
 			<EditorContainerBlock
 				attributes={ attributes }
-				className="wc-block-reviews-by-product"
 				icon={
 					<Icon
 						icon={ comment }
